requirements.txt
streamlit==1.37.1
langchain==0.2.12
langchain-community==0.2.10
langchain-openai==0.1.18
chromadb==0.5.5
pypdf==4.3.1
tiktoken==0.7.0
python-dotenv==1.0.1
